06.開発・テスト/src/backend/board-game-cafe/requirements_all.txt
boto3==1.35.57
botocore==1.35.57
jmespath==1.0.1
pillow==11.1.0
python-dateutil==2.9.0.post0
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
